getting installed games from provider
