+++ backend/src/steam/mod.rs
@@ -2,6 +2,5 @@ pub mod appinfo;
 pub mod command;
 pub mod discover_games;
 pub mod id_lists;
-pub mod installed_games;
 pub mod owned_games;
 pub mod thumbnail;
